#69 added composer.json. psr-2 compatibility. fixes #70
