+++ PHP53to54/Sniffs/Deprecated/FunctionsSniff.php
@@ -1,0 +1,67 @@
+<?php
+/**
+ * Removed Function Aliases Test
+ *
+ * PHP version 5
+ *
+ * @category  PHP
+ * @package   PHP_CodeSniffer
+ * @author    Marcel Eichner // foobugs <marcel.eichner@foobugs.com>
+ * @copyright 2012 foobugs oelke & eichner GbR
+ * @license   BSD http://www.opensource.org/licenses/bsd-license.php
+ * @link      https://github.com/foobugs/PHP53to54
+ * @since     1.0-beta
+ */
+
+namespace PHP53to54\Sniffs\Deprecated;
+
+/**
+ * Deprecated Function Call
+ *
+ * Checks PHP source files for calls to functions that have been removed in
+ * PHP 5.4.
+ *
+ * @category  PHP
+ * @package   PHP_CodeSniffer
+ * @author    Marcel Eichner // foobugs <marcel.eichner@foobugs.com>
+ * @copyright 2012 foobugs oelke & eichner GbR
+ * @license   BSD http://www.opensource.org/licenses/bsd-license.php
+ * @link      https://github.com/foobugs/PHP53to54
+ * @since     1.0-beta
+ */
+class FunctionsSniff extends \Generic_Sniffs_PHP_DeprecatedFunctionsSniff
+{
+    /**
+     * A list of tokenizers this sniff supports.
+     *
+     * @var array
+     */
+    public $supportedTokenizers = array(
+        'PHP',
+    );
+
+    /**
+     * A list of deprecated functions with their alternatives.
+     *
+     * The value is NULL if no alternative exists. IE, the
+     * function should just not be used.
+     *
+     * @var array(string => string|null)
+     */
+    protected $forbiddenFunctions = array(
+        'define_syslog_variables' => null,
+        'import_request_variables' => 'consider using extract',
+        'session_is_registered' => 'use $_SESSION',
+        'session_register' => 'use $_SESSION',
+        'session_unregister' => 'use $_SESSION',
+        'mcrypt_generic_end' => null,
+        'mysql_list_dbs' => null,
+    );
+
+    /**
+     * If true, an error will be thrown; otherwise a warning.
+     *
+     * @var bool
+     */
+    public $error = true;
+}
