+++ composer.json
@@ -1,0 +1,36 @@
+{
+    "name": "foobugs/PHP53to54",
+    "description": "PHP53to54 is a collection of sniffs for PHP_CodeSniffer that check an PHP 5.3 application for PHP 5.4 compatibility.",
+    "type": "library",
+    "keywords": [
+        "phpcs",
+        "standards",
+        "php53",
+        "php54",
+        "compatibility"
+    ],
+    "homepage": "https://github.com/foobugs/PHP53to54",
+    "license": "BSD",
+    "authors": [
+        {
+            "name": "Marcel Eichner"
+        },
+        {
+            "name": "Maik Penz"
+        },
+        {
+            "name": "René Oelke"
+        }
+    ],
+    "support": {
+        "issues": "https://github.com/foobugs/PHP53to54/issues",
+        "source": "https://github.com/foobugs/PHP53to54"
+    },
+    "require": {
+        "php": ">=5.3",
+        "squizlabs/php_codesniffer": "1.4.*"
+    },
+    "autoload": {
+        "psr-0": { "PHP53to54\\": ".\PHP53to54" }
+    }
+}
